+++ LP3/DAA/src/Item.java
@@ -1,8 +1,8 @@
 class Item {
-    float weight;
+    int weight;
     int value;
 
-    Item(int value, float weight) {
+    Item(int value, int weight) {
         this.weight = weight;
         this.value = value;
     }
